fixes resnet - running version
