+++ nupic/research/frameworks/dynamic_sparse/common/utils.py
@@ -38,144 +38,9 @@ from nupic.research.frameworks.dynamic_sparse.common.dataloaders import (
 )
 from nupic.research.frameworks.pytorch.image_transforms import RandomNoise
 from nupic.research.frameworks.pytorch.model_utils import set_random_seed
+from nupic.research.frameworks.pytorch.tiny_imagenet_dataset import TinyImageNet
 
-
-class Dataset:
-    """Loads a dataset.
-    Returns object with a pytorch train and test loader
-    """
-
-    def __init__(self, config=None):
-
-        defaults = dict(
-            dataset_name=None,
-            data_dir=None,
-            batch_size_train=128,
-            batch_size_test=128,
-            stats_mean=None,
-            stats_std=None,
-            augment_images=False,
-            test_noise=False,
-            noise_level=0.1,
-        )
-        defaults.update(config)
-        self.__dict__.update(defaults)
-        self.data_dir = os.path.expanduser(self.data_dir)
-
-        if hasattr(datasets, self.dataset_name):
-            self.load_from_torch_vision()
-        elif self.dataset_name == "PreprocessedGSC":
-            self.load_preprocessed_gsc()
-        else:
-            raise Exception("Dataset {}")
-
-    def load_preprocessed_gsc(self):
-
-        self.train_loader = PreprocessedSpeechDataLoader(
-            self.data_dir,
-            subset="train",
-            batch_sizes=self.batch_size_train,
-            shuffle=True,
-        )
-
-        self.test_loader = PreprocessedSpeechDataLoader(
-            self.data_dir,
-            subset="valid",
-            silence_percentage=0,
-            batch_sizes=self.batch_size_test,
-        )
-
-        if self.test_noise:
-            self.noise_loader = PreprocessedSpeechDataLoader(
-                self.data_dir,
-                subset="test_noise",
-                silence_percentage=0,
-                batch_sizes=self.batch_size_test,
-            )
-        else:
-            self.noise_loader = None
-
-    def load_from_torch_vision(self):
-
-        # special dataloader case
-        if isinstance(self.batch_size_train, Iterable) or isinstance(
-            self.batch_size_test, Iterable
-        ):
-            dataloader_type = VaryingDataLoader
-        else:
-            dataloader_type = DataLoader
-
-        # expand ~
-        self.data_dir = os.path.expanduser(self.data_dir)
-
-        # recover mean and std to normalize dataset
-        if not self.stats_mean or not self.stats_std:
-            tempset = getattr(datasets, self.dataset_name)(
-                root=self.data_dir, train=True, transform=transforms.ToTensor()
-            )
-            if isinstance(tempset.data, np.ndarray):
-                self.stats_mean = (tempset.data.mean() / 255,)
-                self.stats_std = (tempset.data.std() / 255,)
-            else:
-                self.stats_mean = (tempset.data.float().mean().item() / 255,)
-                self.stats_std = (tempset.data.float().std().item() / 255,)
-            del tempset
-
-        # set up transformations
-        transform = transforms.Compose(
-            [
-                transforms.ToTensor(),
-                transforms.Normalize(self.stats_mean, self.stats_std),
-            ]
-        )
-        # set up augment transforms for training
-        if not self.augment_images:
-            aug_transform = transform
-        else:
-            aug_transform = transforms.Compose(
-                [
-                    transforms.RandomCrop(32, padding=4),
-                    transforms.RandomHorizontalFlip(),
-                    transforms.ToTensor(),
-                    transforms.Normalize(self.stats_mean, self.stats_std),
-                ]
-            )
-
-        # load train set
-        train_set = getattr(datasets, self.dataset_name)(
-            root=self.data_dir, train=True, transform=aug_transform
-        )
-        self.train_loader = dataloader_type(
-            dataset=train_set, batch_size=self.batch_size_train, shuffle=True
-        )
-
-        # load test set
-        test_set = getattr(datasets, self.dataset_name)(
-            root=self.data_dir, train=False, transform=transform
-        )
-        self.test_loader = dataloader_type(
-            dataset=test_set, batch_size=self.batch_size_test, shuffle=False
-        )
-
-        # noise dataset
-        if self.test_noise:
-            noise = self.noise_level
-            noise_transform = transforms.Compose(
-                [
-                    transforms.ToTensor(),
-                    transforms.Normalize(self.stats_mean, self.stats_std),
-                    RandomNoise(
-                        noise, high_value=0.5 + 2 * 0.20, low_value=0.5 - 2 * 0.2
-                    ),
-                ]
-            )
-            noise_set = getattr(datasets, self.dataset_name)(
-                root=self.data_dir, train=False, transform=noise_transform
-            )
-            self.noise_loader = dataloader_type(
-                dataset=noise_set, batch_size=self.batch_size_test, shuffle=False
-            )
-
+from .datasets import Dataset
 
 class Trainable(tune.Trainable):
     """ray.tune trainable generic class Adaptable to any pytorch module."""
@@ -200,7 +65,6 @@ class Trainable(tune.Trainable):
     def _restore(self, checkpoint):
         self.model.restore(checkpoint)
 
-
 def download_dataset(config):
     """Pre-downloads dataset.
     Required to avoid multiple simultaneous attempts to download same
@@ -212,12 +76,10 @@ def download_dataset(config):
             download=True, root=os.path.expanduser(config["data_dir"])
         )
 
-
 def new_experiment(base_config, new_config):
     modified_config = deepcopy(base_config)
     modified_config.update(new_config)
     return modified_config
-
 
 @ray.remote
 def run_experiment(name, trainable, exp_config, tune_config):
@@ -234,7 +96,6 @@ def run_experiment(name, trainable, exp_config, tune_config):
     # tune_config["name"] = name
     tune_config["config"] = exp_config
     tune.run(Trainable, **tune_config)
-
 
 def init_ray():
 
@@ -264,7 +125,6 @@ def init_ray():
             t, serializer=serializer, deserializer=deserializer
         )
 
-
 def run_ray(tune_config, exp_config, fix_seed=False):
 
     # update config
@@ -275,6 +135,17 @@ def run_ray(tune_config, exp_config, fix_seed=False):
     if not torch.cuda.is_available():
         tune_config["config"]["device"] = "cpu"
         tune_config["resources_per_trial"] = {"cpu": 1}
+
+    # move epochs to tune_config, to keep track
+    if 'stop' not in tune_config:
+        if 'epochs' in exp_config:
+            tune_config['stop'] = {"training_iteration": exp_config['epochs']}
+
+    # expand path in dir
+    if 'local_dir' in tune_config:
+        tune_config['local_dir']=os.path.expanduser(tune_config['local_dir'])
+    else:
+        tune_config['local_dir']=os.path.expanduser("~/nta/results")
 
     # init ray
     ray.init(load_code_from_local=True)
@@ -309,7 +180,6 @@ def run_ray(tune_config, exp_config, fix_seed=False):
         set_random_seed(32)
 
     tune.run(Trainable, **tune_config)
-
 
 def run_ray_many(tune_config, exp_config, experiments, fix_seed=False):
