+++ nupic/research/frameworks/dynamic_sparse/networks/__init__.py
@@ -3,4 +3,5 @@ from .hebbian import *
 from .layers import *
 from .main import *
 from .utils import *
-from .wideresnet import *
+# from .wideresnet import *
+from .resnet import *
